Replace size_t by unsigned short for entity ids

+++ src/client/systems/bonus/System+MoveBonus.cpp
@@ -10,7 +10,7 @@ namespace ECS {
         auto &world = Core::World::getInstance();
         const auto size = aType.size();
 
-        for (size_t idx = 0; idx < size; idx++) {
+        for (unsigned short idx = 0; idx < size; idx++) {
             if (!aPos[idx].has_value() || !aSpeed[idx].has_value() || !aType[idx].has_value()) {
                 continue;
             }
