+++ src/client/systems/enemy/System+MoveEnemy.cpp
@@ -11,7 +11,7 @@ namespace ECS {
         auto &display = SFMLDisplayClass::getInstance();
         const auto size = aPos.size();
 
-        for (size_t idx = 0; idx < size; idx++) {
+        for (unsigned short idx = 0; idx < size; idx++) {
             if (!aPos[idx].has_value() || !aSpeed[idx].has_value() || !aType[idx].has_value()) {
                 continue;
             }
