+++ src/client/systems/graphic/System+CreateServerFullErrorMessage.cpp
@@ -18,7 +18,7 @@ namespace ECS {
         const auto size = events.size();
         std::vector<size_t> toRemove;
 
-        for (size_t i = 0; i < size; i++) {
+        for (unsigned short i = 0; i < size; i++) {
             auto &gameEvent = events[i];
 
             if (gameEvent.getType() != RType::ClientEventType::SERVER_FULL) {
